refactor: move AgendaRepository to agenda/data
