+++ shared/src/commonMain/kotlin/com/example/tasky/agenda/data/AgendaRepository.kt
@@ -1,11 +1,10 @@
-package com.example.tasky.agenda.domain
+package com.example.tasky.agenda.data
 
-import com.example.tasky.agenda.data.AgendaDataSource
-import com.example.tasky.agenda.data.AgendaLocalDataSource
 import com.example.tasky.agenda.data.model.CreateEventBody
 import com.example.tasky.agenda.data.model.RemoteReminder
 import com.example.tasky.agenda.data.model.RemoteTask
 import com.example.tasky.agenda.data.model.UpdateEventBody
+import com.example.tasky.agenda.domain.IAgendaRepository
 import com.example.tasky.agenda.domain.model.AgendaItem
 import com.example.tasky.agenda.domain.model.Attendee
 import com.example.tasky.agenda.domain.model.Event
@@ -20,43 +19,6 @@ import com.example.tasky.database.model.isDelete
 import com.example.tasky.login.domain.manager.SessionManager
 import dev.tmapps.konnection.Konnection
 import kotlinx.serialization.json.Json
-
-interface IAgendaRepository {
-    suspend fun getAgenda(timeStamp: Long): ResultWrapper<List<AgendaItem>, BaseError>
-
-    suspend fun deleteAgenda(agendaItem: AgendaItem): ResultWrapper<Unit, BaseError>
-
-    suspend fun updateTask(task: Task): ResultWrapper<Unit, BaseError>
-
-    suspend fun updateReminder(reminder: Reminder): ResultWrapper<Unit, BaseError>
-
-    suspend fun updateEvent(
-        event: Event,
-        deletedPhotoKeys: List<String>,
-        isGoing: Boolean,
-        photos: List<ByteArray>,
-    ): ResultWrapper<Event, BaseError>
-
-    suspend fun createTask(task: Task): ResultWrapper<Unit, BaseError>
-
-    suspend fun createReminder(reminder: Reminder): ResultWrapper<Unit, BaseError>
-
-    suspend fun createEvent(event: Event): ResultWrapper<Event, BaseError>
-
-    suspend fun getTask(taskId: String): ResultWrapper<Task, BaseError>
-
-    suspend fun getEvent(eventId: String): ResultWrapper<Event, BaseError>
-
-    suspend fun getReminder(reminderId: String): ResultWrapper<Reminder, BaseError>
-
-    suspend fun getAttendee(
-        email: String,
-        eventId: String,
-        from: Long,
-    ): ResultWrapper<Attendee?, BaseError>
-
-    suspend fun syncAgenda()
-}
 
 class AgendaRepository(
     private val agendaDataSource: AgendaDataSource = AgendaDataSource(),
