+++ shared/src/commonMain/kotlin/com/example/tasky/agenda/domain/IAgendaRepository.kt
@@ -1,0 +1,46 @@
+package com.example.tasky.agenda.domain
+
+import com.example.tasky.agenda.domain.model.AgendaItem
+import com.example.tasky.agenda.domain.model.Attendee
+import com.example.tasky.agenda.domain.model.Event
+import com.example.tasky.agenda.domain.model.Reminder
+import com.example.tasky.agenda.domain.model.Task
+import com.example.tasky.common.data.model.BaseError
+import com.example.tasky.common.domain.model.ResultWrapper
+
+interface IAgendaRepository {
+    suspend fun getAgenda(timeStamp: Long): ResultWrapper<List<AgendaItem>, BaseError>
+
+    suspend fun deleteAgenda(agendaItem: AgendaItem): ResultWrapper<Unit, BaseError>
+
+    suspend fun updateTask(task: Task): ResultWrapper<Unit, BaseError>
+
+    suspend fun updateReminder(reminder: Reminder): ResultWrapper<Unit, BaseError>
+
+    suspend fun updateEvent(
+        event: Event,
+        deletedPhotoKeys: List<String>,
+        isGoing: Boolean,
+        photos: List<ByteArray>,
+    ): ResultWrapper<Event, BaseError>
+
+    suspend fun createTask(task: Task): ResultWrapper<Unit, BaseError>
+
+    suspend fun createReminder(reminder: Reminder): ResultWrapper<Unit, BaseError>
+
+    suspend fun createEvent(event: Event): ResultWrapper<Event, BaseError>
+
+    suspend fun getTask(taskId: String): ResultWrapper<Task, BaseError>
+
+    suspend fun getEvent(eventId: String): ResultWrapper<Event, BaseError>
+
+    suspend fun getReminder(reminderId: String): ResultWrapper<Reminder, BaseError>
+
+    suspend fun getAttendee(
+        email: String,
+        eventId: String,
+        from: Long,
+    ): ResultWrapper<Attendee?, BaseError>
+
+    suspend fun syncAgenda()
+}
